// src/model/enums/CardCodesTypes.tsx
export enum CardCodesTypes {
    Visa = 1, // MC
    MasterCard = 2,
    AmericanExpress = 3,
    Discover = 4,
    HiperCard = 5,
    DinnersClub = 6,
    JCB = 7,
    Electron = 8,
    EuroCard = 9,
    EnRouteCard = 10,
    Laser = 11,
    Maestro = 12,
    Solo = 13,
    PayPal = 14,
    Aura = 15,
    Liberate = 16,
    Unknown = 100,
    Interbank = 101,
    UATP = 102,
    Access = 103,
    CarteBleue = 104
}